REF: Moves the validation logic to a middleware
